consumer/requirements.txt
kafka-python==2.0.2
ipaddress==1.0.23
pytest==7.4.2